New method for retrieving items (terms) from the DB.
We avoid rendering the result of XQuery queries as jinja templates
because it's *very* expensive and badly affects overall performance.

Now we return a sequence of strings that will be parsed into actual Term
objects which will be passed into the templates for _pleasant_
templating and rendering.

This commit only covers the glossary and search modules, but the terms
module would need a similar change too.